netconvert: read floating point values from --heightmap.geotiff instead of integers for more precision

+++ src/netbuild/NBHeightMapper.h
@@ -124,7 +124,7 @@ public:
 private:
 
     struct RasterData {
-        int16_t* raster;
+        float* raster;
         Boundary boundary;
         int xSize;
         int ySize;
